Workspace refactoring event + some changes/java doc
